refactor: split claim wizard for reviewability
